requirements.txt
scikit-learn==0.23.2
pillow==8.0.1
scipy
flask
flask_restful